+++ Packages/OpenApiCodeGen/Editor/Core/GenerateProvider.cs
@@ -1,9 +1,0 @@
-namespace ReBeat.OpenApiCodeGen.Core
-{
-    public enum GenerateProvider
-    {
-        OpenApi,
-        // Swagger,
-    }
-
-}