+++ Packages/OpenApiCodeGen/Editor/Domain/Setting/UserSetting.cs
@@ -1,0 +1,12 @@
+﻿namespace ReBeat.OpenApiCodeGen.Core
+{
+    internal class UserSetting
+    {
+        public string DockerPath { get; private set; }
+
+        public UserSetting(string dockerPath)
+        {
+            this.DockerPath = dockerPath;
+        }
+    }
+}