+++ Packages/OpenApiCodeGen/Editor/Lib/OpenApi/OpenApiGenerateTool.cs
@@ -1,7 +1,0 @@
-namespace ReBeat.OpenApiCodeGen.Lib
-{
-    public enum OpenApiGenerateTool
-    {
-        Docker
-    }
-}