+++ Packages/OpenApiCodeGen/CHANGELOG.md
@@ -1,9 +1,21 @@
-# Change Log
+﻿# Change Log
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## unreleased
+
+## Feature
+
+- Display Generated Result on Generate Menu
+
+### Fixed
+
+- Change Setting Menu Layout
+- Independent From SwaggerGen
+- Available to Use Relative Path in Output Path(from project root)
 
 ## [0.2.2]